Update EventsViewController.swift
So what's been done is that a tableview is made, but cells only display one line. Also can't figure out how to retrieve user specific data (i had one account with events, another without, and the empty user came up with the same results as the one with events). Added swipe to delete func that swipes but doesn't delete.